+++ src/world/Server/Packets/Opcode.h
@@ -22,15 +22,15 @@ This file is released under the MIT license. See README-MIT for more information
 
 struct OpcodeTable
 {
-    uint32_t opcode;
+    uint16_t opcode;
     std::string name;
 };
 
 extern OpcodeTable opcodeNameArray[];
 
-static inline std::string getOpcodeName(uint32_t opcode)
+static inline std::string getOpcodeName(uint16_t opcode)
 {
-    for (uint32_t i = 0; i < NUM_MSG_TYPES; ++i)
+    for (uint16_t i = 0; i < NUM_MSG_TYPES; ++i)
     {
         if (opcodeNameArray[i].opcode == opcode)
             return opcodeNameArray[i].name;
